+++ mtpy/modeling/ModEM/__init__.py
@@ -3,6 +3,7 @@ from .station import Stations
 from .data import Data
 from .model import Model
 from .residual import Residual
-from .controlinv import ControlInv
+from .control_inv import ControlInv
+from .control_fwd import ControlFwd
 
-__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual', 'ControlInv']
+__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual', 'ControlInv', 'ControlFwd']
